\name{as.matrix.pdist}
\alias{as.matrix.pdist}
\title{Convert pdist to a distance matrix}
\usage{
  \method{as.matrix}{pdist} (x, ...)
}
\arguments{
  \item{x}{an instance of class pdist}

  \item{...}{additional arguments for as.matrix signature;
  currently not used}
}
\description{
  pdist objects are numeric arrays, converts this to matrix
}

