\name{pdist-package}
\alias{pdist-package}
\docType{package}
\title{
Distances between Observations for a Partitioned Matrix
}
\description{
Computes the distances between rows of a matrix block X
and rows of a matrix block Y.  Much faster than dist for certain
use cases.
}
\details{
\tabular{ll}{
Package: \tab pdist\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2012-12-28\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Jeffrey Wong <jeff.ct.wong@gmail.com>
}
\keyword{ package }
