\name{pdist-package}
\alias{pdist-package}
\docType{package}
\title{
Distances between Observations for a Partitioned Matrix
}
\description{
Computes the distances between rows of a matrix block X
and rows of a matrix block Y.
}
\details{
\tabular{ll}{
Package: \tab pdist\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-18\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Jeffrey Wong <jeff.ct.wong@stanford.edu>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
